Move listening module to user_interface package
